requirements.txt
pyannote.audio
hydra_core
hydra_colorlog
julius
numpy
pystoi
git+https://github.com/ludlows/python-pesq#egg=pesq
six
sounddevice
torchaudio
torch
speechbrain
denoiser
mir_eval
pytorch-minimize
